@@ -1,34 +1,29 @@
 {
-    "name": "hacklabsdev/laravel-trends",
-    "description": "Laravel Trends provides a lightweight trending system to your application.",
-    "type": "library",
-    "license": "MIT",
-    "authors": [
-        {
-            "name": "Felipe Martins",
-            "email": "felipe@hacklabs.com.br"
-        }
-    ],
-    "minimum-stability": "dev",
-    "autoload": {
-        "psr-4": {
-            "Hacklabs\\Trends\\": "src/"
-        }
-    },
-    "extra": {
-        "laravel": {
-            "providers": [
-                "Hacklabs\\Trends\\TrendsServiceProvider"
-            ],
-            "aliases": {
-                "LaravelRating": "Hacklabs\\Trends\\TrendsFacade"
-            }
-        }
-    },
-    "require": {
-        "php": "^7.2|^8.0",
-        "illuminate/contracts": "^6.9|^7.0|^8.0",
-        "illuminate/database": "^6.9|^7.0|^8.0",
-        "illuminate/support": "^6.9|^7.0|^8.0"
-    }
+  "name": "hawaaworld/laravel-trends",
+  "description": "Laravel Trends provides a lightweight trending system to your application.",
+  "type": "library",
+  "license": "MIT",
+  "authors": [
+	{
+	  "name": "Mohammad ALTAWEEL",
+	  "email": "hello@mohammad.tech"
+	}
+  ],
+  "minimum-stability": "dev",
+  "autoload": {
+	"psr-4": {
+	  "Hawaaworld\\Trends\\": "src/"
+	}
+  },
+  "extra": {
+	"laravel": {
+	  "providers": [
+		"Hawaaworld\\Trends\\TrendsServiceProvider"
+	  ]
+	}
+  },
+  "require": {
+	"php": "^8.0",
+	"laravel/framework": "^9.0"
+  }
 }
