@@ -3,15 +3,15 @@
 return [
     /*
     |--------------------------------------------------------------------------
-    | Energy Decay Time
+    | Trends Loss Time
     |--------------------------------------------------------------------------
     |
-    | Defines how long your energy starts to decay in hours. The energy decays 
-    | three times until it returns to 0. For example if value is set to 24 hours 
-    | it would take 3 days to energy return to zero.
+    | Defines how long your energy starts to loss in minutes. The energy
+    | losses three times until it returns to 0. For example if value is
+    | set to 60 minutes it would take 3 hours to energy return to zero.
     |
     */
-    'energy_decay' => env('TRENDS_DECAY_TIME', 24),
+    'loss_time' => env('TRENDS_LOSS_TIME', 60),
 
     /*
     |--------------------------------------------------------------------------
@@ -23,7 +23,17 @@ return [
     |
     */
     'ip_blacklist' => [
-        // 
-    ]
+        //
+    ],
 
+    /*
+    |--------------------------------------------------------------------------
+    | Truncate Zero Energy Trends
+    |--------------------------------------------------------------------------
+    |
+    | Defines if the trends should be truncated. If set to true,
+    | the trends with the zero energy will be deleted immediately.
+    |
+    */
+    'truncate' => true,
 ];
