@@ -1,14 +1,16 @@
 <?php
 
-namespace Hacklabs\Trends\Models;
+namespace Hawaaworld\Trends\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\MorphTo;
 
 class Energy extends Model
 {
     protected $fillable = ['amount'];
 
-    public function subject() {
+    public function subject(): MorphTo
+    {
         return $this->morphTo();
     }
 }
