@@ -1,17 +1,31 @@
 <?php
 
-namespace Hacklabs\Trends;
+namespace Hawaaworld\Trends;
+
+use Hawaaworld\Trends\Contracts\Energy as EnergyContract;
+use Hawaaworld\Trends\Models\Energy;
+use Illuminate\Contracts\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Database\Eloquent\Model;
 
 class Trends
 {
-    public function add($model, $amount)
+    public function add(EnergyContract $model, float $amount = 1.0): void
     {
-        return $model->addEnergy($amount);
+        $model->addEnergy($amount);
     }
 
-    public function getEnergy($model)
+    public function getEnergy($model): float
     {
-        $energy = $model->energy;
-        return $energy !== null ? (float) $energy->amount : null;
+        return (float) $model->energy->amount;
+    }
+
+    public function top(int $limit = 10, string $model = null): Collection
+    {
+        return Energy::query()
+            ->when($model, fn (Builder $query) => $query->where('energiser_type', $model))
+            ->latest('amount')
+            ->take($limit)
+            ->get();
     }
 }
